[FEAT] #36 - MonthlyScreen과 MonthlyBody 분리
