+++ feature/monthly/src/main/java/com/feature/monthly/MonthlyScreen.kt
@@ -2,31 +2,21 @@ package com.feature.monthly
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.material.LocalContentColor
-import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Brush
 import androidx.hilt.navigation.compose.hiltViewModel
+import com.core.designsystem.R
 import com.core.designsystem.components.HarooHeader
-import com.core.designsystem.components.HarooRadioButton
-import com.core.designsystem.components.HarooSurface
 import com.core.designsystem.theme.HarooTheme
 import com.core.designsystem.util.getString
-import com.core.model.feature.PostUiModel
-import com.core.ui.post.GridPostItem
-import com.core.ui.post.LinearPostItem
 import com.core.ui.toolbar.CollapsingToolbar
 import com.feature.monthly.ui.Dimens
+import com.feature.monthly.ui.MonthlyBody
 import com.feature.monthly.ui.MonthlyHeader
-import java.time.LocalDate
-import java.time.YearMonth
-import com.core.designsystem.R
 
 @Composable
 fun MonthlyScreen(
@@ -78,76 +68,5 @@ fun MonthlyScreen(
                 onRemovePost = monthlyScreenStateHolder::removePost
             )
         }
-    }
-}
-
-@Composable
-fun MonthlyBody(
-    lazyListState: LazyListState,
-    groupedPosts: Map<LocalDate, PostUiModel>,
-    listType: Boolean,
-    date: YearMonth,
-    dateCount: Int,
-    onChangeListType: () -> Unit,
-    onRemovePost: (PostUiModel) -> Unit
-) {
-    HarooSurface(
-        modifier = Modifier.fillMaxWidth(),
-        shape = MaterialTheme.shapes.large,
-        alpha = 0.08f
-    ) {
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            HarooRadioButton(
-                modifier = Modifier.padding(Dimens.paddingRadioBtn),
-                selected = listType,
-                onSelected = onChangeListType
-            )
-            LazyColumn(
-                state = lazyListState,
-                contentPadding = PaddingValues(horizontal = Dimens.postListHorizontalPadding)
-            ) {
-                items(count = dateCount) {
-                    val day = date.atDay(it + 1)
-                    MonthlyPostItem(
-                        isFirstItem = it == 0,
-                        isLastItem = it == dateCount - 1,
-                        date = day,
-                        post = groupedPosts[day],
-                        listType = listType,
-                        onRemovePost = onRemovePost
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun MonthlyPostItem(
-    isFirstItem: Boolean,
-    isLastItem: Boolean,
-    date: LocalDate,
-    post: PostUiModel?,
-    listType: Boolean,
-    onRemovePost: (PostUiModel) -> Unit
-) {
-    if (listType) {
-        GridPostItem(
-            isFirstItem = isFirstItem,
-            isLastItem = isLastItem,
-            date = date,
-            post = post,
-            onRemovePost = onRemovePost
-        )
-    } else {
-        LinearPostItem(
-            isFirstItem = isFirstItem,
-            isLastItem = isLastItem,
-            date = date,
-            post = post,
-            onRemovePost = onRemovePost
-        )
     }
 }
